requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.0.3
frozenlist==1.5.0
greenlet==3.1.1
groq==0.12.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.6
langchain-core==0.3.16
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.142
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.11
packaging==24.2
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-decouple==3.8
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.1
youtube-transcript-api==0.6.2
yt-dlp==2024.11.4
